+++ thoth-app/src/component/affiliations_form.rs
@@ -14,7 +14,7 @@ use crate::agent::notification_bus::NotificationBus;
 use crate::agent::notification_bus::NotificationDispatcher;
 use crate::agent::notification_bus::NotificationStatus;
 use crate::agent::notification_bus::Request;
-use crate::component::utils::FormInstitutionSelect;
+use crate::component::institution_select::InstitutionSelectComponent;
 use crate::component::utils::FormNumberInput;
 use crate::component::utils::FormTextInput;
 use crate::models::affiliation::affiliations_query::AffiliationsRequest;
@@ -37,12 +37,6 @@ use crate::models::affiliation::update_affiliation_mutation::PushUpdateAffiliati
 use crate::models::affiliation::update_affiliation_mutation::UpdateAffiliationRequest;
 use crate::models::affiliation::update_affiliation_mutation::UpdateAffiliationRequestBody;
 use crate::models::affiliation::update_affiliation_mutation::Variables as UpdateVariables;
-use crate::models::institution::institutions_query::FetchActionInstitutions;
-use crate::models::institution::institutions_query::FetchInstitutions;
-use crate::models::institution::institutions_query::InstitutionsRequest;
-use crate::models::institution::institutions_query::InstitutionsRequestBody;
-use crate::models::institution::institutions_query::Variables as SearchVariables;
-use crate::models::Dropdown;
 use crate::string::CANCEL_BUTTON;
 use crate::string::EDIT_BUTTON;
 use crate::string::REMOVE_BUTTON;
@@ -52,32 +46,20 @@ use super::ToOption;
 
 pub struct AffiliationsFormComponent {
     fetch_affiliations: FetchAffiliations,
-    data: AffiliationsFormData,
+    affiliations: Option<Vec<AffiliationWithInstitution>>,
     affiliation: AffiliationWithInstitution,
     show_modal_form: bool,
     in_edit_mode: bool,
-    show_results: bool,
-    fetch_institutions: FetchInstitutions,
     create_affiliation: PushCreateAffiliation,
     delete_affiliation: PushDeleteAffiliation,
     update_affiliation: PushUpdateAffiliation,
     notification_bus: NotificationDispatcher,
 }
 
-#[derive(Default)]
-struct AffiliationsFormData {
-    institutions: Vec<Institution>,
-    affiliations: Option<Vec<AffiliationWithInstitution>>,
-}
-
 pub enum Msg {
     ToggleModalFormDisplay(bool, Option<AffiliationWithInstitution>),
     SetAffiliationsFetchState(FetchActionAffiliations),
     GetAffiliations,
-    SetInstitutionsFetchState(FetchActionInstitutions),
-    GetInstitutions,
-    ToggleSearchResultDisplay(bool),
-    SearchInstitution(String),
     SetAffiliationCreateState(PushActionCreateAffiliation),
     CreateAffiliation,
     SetAffiliationUpdateState(PushActionUpdateAffiliation),
@@ -85,7 +67,7 @@ pub enum Msg {
     SetAffiliationDeleteState(PushActionDeleteAffiliation),
     DeleteAffiliation(Uuid),
     AddAffiliation(Institution),
-    ChangeInstitution(Uuid),
+    ChangeInstitution(Institution),
     ChangePosition(String),
     ChangeOrdinal(String),
 }
@@ -101,28 +83,23 @@ impl Component for AffiliationsFormComponent {
 
     fn create(ctx: &Context<Self>) -> Self {
         let fetch_affiliations = Default::default();
-        let data: AffiliationsFormData = Default::default();
+        let affiliations: Option<Vec<AffiliationWithInstitution>> = Default::default();
         let affiliation: AffiliationWithInstitution = Default::default();
         let show_modal_form = false;
         let in_edit_mode = false;
-        let show_results = false;
-        let fetch_institutions = Default::default();
         let create_affiliation = Default::default();
         let delete_affiliation = Default::default();
         let update_affiliation = Default::default();
         let notification_bus = NotificationBus::dispatcher();
 
         ctx.link().send_message(Msg::GetAffiliations);
-        ctx.link().send_message(Msg::GetInstitutions);
 
         AffiliationsFormComponent {
             fetch_affiliations,
-            data,
+            affiliations,
             affiliation,
             show_modal_form,
             in_edit_mode,
-            show_results,
-            fetch_institutions,
             create_affiliation,
             delete_affiliation,
             update_affiliation,
@@ -136,16 +113,6 @@ impl Component for AffiliationsFormComponent {
                 self.show_modal_form = show_form;
                 self.in_edit_mode = a.is_some();
                 if show_form {
-                    let body = InstitutionsRequestBody {
-                        variables: SearchVariables {
-                            limit: Some(9999),
-                            ..Default::default()
-                        },
-                        ..Default::default()
-                    };
-                    let request = InstitutionsRequest { body };
-                    self.fetch_institutions = Fetch::new(request);
-                    ctx.link().send_message(Msg::GetInstitutions);
                     if let Some(affiliation) = a {
                         // Editing existing affiliation: load its current values.
                         self.affiliation = affiliation;
@@ -155,7 +122,7 @@ impl Component for AffiliationsFormComponent {
             }
             Msg::SetAffiliationsFetchState(fetch_state) => {
                 self.fetch_affiliations.apply(fetch_state);
-                self.data.affiliations = match self.fetch_affiliations.as_ref().state() {
+                self.affiliations = match self.fetch_affiliations.as_ref().state() {
                     FetchState::NotFetching(_) => None,
                     FetchState::Fetching(_) => None,
                     FetchState::Fetched(body) => match &body.data.contribution {
@@ -184,25 +151,6 @@ impl Component for AffiliationsFormComponent {
                     .send_message(Msg::SetAffiliationsFetchState(FetchAction::Fetching));
                 false
             }
-            Msg::SetInstitutionsFetchState(fetch_state) => {
-                self.fetch_institutions.apply(fetch_state);
-                self.data.institutions = match self.fetch_institutions.as_ref().state() {
-                    FetchState::NotFetching(_) => vec![],
-                    FetchState::Fetching(_) => vec![],
-                    FetchState::Fetched(body) => body.data.institutions.clone(),
-                    FetchState::Failed(_, _err) => vec![],
-                };
-                true
-            }
-            Msg::GetInstitutions => {
-                ctx.link().send_future(
-                    self.fetch_institutions
-                        .fetch(Msg::SetInstitutionsFetchState),
-                );
-                ctx.link()
-                    .send_message(Msg::SetInstitutionsFetchState(FetchAction::Fetching));
-                false
-            }
             Msg::SetAffiliationCreateState(fetch_state) => {
                 self.create_affiliation.apply(fetch_state);
                 match self.create_affiliation.clone().state() {
@@ -212,9 +160,9 @@ impl Component for AffiliationsFormComponent {
                         Some(a) => {
                             let affiliation = a.clone();
                             let mut affiliations: Vec<AffiliationWithInstitution> =
-                                self.data.affiliations.clone().unwrap_or_default();
+                                self.affiliations.clone().unwrap_or_default();
                             affiliations.push(affiliation);
-                            self.data.affiliations = Some(affiliations);
+                            self.affiliations = Some(affiliations);
                             ctx.link()
                                 .send_message(Msg::ToggleModalFormDisplay(false, None));
                             true
@@ -268,13 +216,13 @@ impl Component for AffiliationsFormComponent {
                     FetchState::Fetched(body) => match &body.data.update_affiliation {
                         Some(a) => {
                             let mut affiliations: Vec<AffiliationWithInstitution> =
-                                self.data.affiliations.clone().unwrap_or_default();
+                                self.affiliations.clone().unwrap_or_default();
                             if let Some(affiliation) = affiliations
                                 .iter_mut()
                                 .find(|af| af.affiliation_id == a.affiliation_id)
                             {
                                 *affiliation = a.clone();
-                                self.data.affiliations = Some(affiliations);
+                                self.affiliations = Some(affiliations);
                             } else {
                                 // This should not be possible: the updated affiliation returned from the
                                 // database does not match any of the locally-stored affiliation data.
@@ -338,14 +286,13 @@ impl Component for AffiliationsFormComponent {
                     FetchState::Fetched(body) => match &body.data.delete_affiliation {
                         Some(affiliation) => {
                             let to_keep: Vec<AffiliationWithInstitution> = self
-                                .data
                                 .affiliations
                                 .clone()
                                 .unwrap_or_default()
                                 .into_iter()
                                 .filter(|a| a.affiliation_id != affiliation.affiliation_id)
                                 .collect();
-                            self.data.affiliations = Some(to_keep);
+                            self.affiliations = Some(to_keep);
                             true
                         }
                         None => {
@@ -387,40 +334,9 @@ impl Component for AffiliationsFormComponent {
                     .send_message(Msg::ToggleModalFormDisplay(true, None));
                 true
             }
-            Msg::ToggleSearchResultDisplay(value) => {
-                self.show_results = value;
-                true
-            }
-            Msg::SearchInstitution(value) => {
-                let body = InstitutionsRequestBody {
-                    variables: SearchVariables {
-                        filter: Some(value),
-                        limit: Some(9999),
-                        ..Default::default()
-                    },
-                    ..Default::default()
-                };
-                let request = InstitutionsRequest { body };
-                self.fetch_institutions = Fetch::new(request);
-                ctx.link().send_message(Msg::GetInstitutions);
-                false
-            }
-            Msg::ChangeInstitution(institution_id) => {
-                // ID may be nil if placeholder option was selected.
-                // Reset institution anyway, to keep display/underlying values in sync.
-                self.affiliation.institution_id.neq_assign(institution_id);
-                // we already have the full list of institutions
-                if let Some(institution) = self
-                    .data
-                    .institutions
-                    .iter()
-                    .find(|i| i.institution_id == institution_id)
-                {
-                    self.affiliation.institution.neq_assign(institution.clone());
-                } else {
-                    // Institution not found: clear existing selection
-                    self.affiliation.institution.neq_assign(Default::default());
-                }
+            Msg::ChangeInstitution(institution) => {
+                self.affiliation.institution_id = institution.institution_id;
+                self.affiliation.institution = institution;
                 true
             }
             Msg::ChangePosition(val) => self.affiliation.position.neq_assign(val.to_opt_string()),
@@ -441,11 +357,12 @@ impl Component for AffiliationsFormComponent {
         // Ensure the form has a unique ID, as there may be multiple copies of
         // the form on the same parent page, and ID clashes can lead to bugs
         let form_id = format!("affiliations-form-{}", ctx.props().contribution_id);
-        let affiliations = self.data.affiliations.clone().unwrap_or_default();
+        let affiliations = self.affiliations.clone().unwrap_or_default();
         let close_modal = ctx.link().callback(|e: MouseEvent| {
             e.prevent_default();
             Msg::ToggleModalFormDisplay(false, None)
         });
+
         html! {
             <div class="field">
                 <div class={ self.modal_form_status() }>
@@ -461,15 +378,13 @@ impl Component for AffiliationsFormComponent {
                         </header>
                         <section class="modal-card-body">
                             <form id={form_id.clone()} onsubmit={ self.modal_form_action(ctx) }>
-                                <FormInstitutionSelect
-                                    label = "Institution"
-                                    value={ self.affiliation.institution_id }
-                                    data={ self.data.institutions.clone() }
-                                    onchange={ ctx.link().callback(|e: Event|
-                                        Msg::ChangeInstitution(Uuid::parse_str(&e.to_value()).unwrap_or_default())
-                                    ) }
-                                    required = true
-                                />
+                                <div class="field">
+                                    <label class="label">{ "Institution" }</label>
+                                    <div class="control is-expanded">
+                                        {&self.affiliation.institution}
+                                    </div>
+                                </div>
+                                <InstitutionSelectComponent callback={ctx.link().callback(Msg::ChangeInstitution)} />
                                 <FormTextInput
                                     label="Position"
                                     value={ self.affiliation.position.clone().unwrap_or_default() }
@@ -521,41 +436,7 @@ impl Component for AffiliationsFormComponent {
                     <tbody>
                         {for affiliations.iter().map(|a| self.render_affiliation(ctx, a))}
                         <tr class="row">
-                            <div class={ self.search_dropdown_status() } style="width: 100%">
-                                <div class="dropdown-trigger" style="width: 100%">
-                                    <div class="field">
-                                        <p class="control is-expanded has-icons-left">
-                                            <input
-                                                class="input"
-                                                type="search"
-                                                placeholder="Search Institution"
-                                                aria-haspopup="true"
-                                                aria-controls="institutions-menu"
-                                                oninput={ ctx.link().callback(|e: InputEvent| Msg::SearchInstitution(e.to_value())) }
-                                                onfocus={ ctx.link().callback(|_| Msg::ToggleSearchResultDisplay(true)) }
-                                                onblur={ ctx.link().callback(|_| Msg::ToggleSearchResultDisplay(false)) }
-                                            />
-                                            <span class="icon is-left">
-                                                <i class="fas fa-search" aria-hidden="true"></i>
-                                            </span>
-                                        </p>
-                                    </div>
-                                </div>
-                                <div class="dropdown-menu" id="institutions-menu" role="menu">
-                                    <div class="dropdown-content">
-                                        {
-                                            for self.data.institutions.iter().map(|i| {
-                                                let institution = i.clone();
-                                                i.as_dropdown_item(
-                                                    ctx.link().callback(move |_| {
-                                                        Msg::AddAffiliation(institution.clone())
-                                                    })
-                                                )
-                                            })
-                                        }
-                                    </div>
-                                </div>
-                            </div>
+                            <InstitutionSelectComponent callback={ctx.link().callback(Msg::AddAffiliation)} />
                         </tr>
                     </tbody>
                 </table>
@@ -596,13 +477,6 @@ impl AffiliationsFormComponent {
                 e.prevent_default();
                 Msg::CreateAffiliation
             }),
-        }
-    }
-
-    fn search_dropdown_status(&self) -> String {
-        match self.show_results {
-            true => "dropdown is-active".to_string(),
-            false => "dropdown".to_string(),
         }
     }