+++ thoth-app/src/models/work_relation/create_work_relation_mutation.rs
@@ -33,6 +33,7 @@ const CREATE_WORK_RELATION_MUTATION: &str = "
                 imprintId
                 createdAt
                 updatedAt
+                updatedAtWithRelations
             }
         }
     }